looking for ways to modify and optimize fragmentation tree creation, also fixed build warnings.

+++ src/AnnotateMs2/AnnotateMs2.h
@@ -10,7 +10,7 @@
 
 class AnnotateMs2 {
 public:
-    AnnotateMs2(const int minPeaks):minPeaks(minPeaks) {}
+    AnnotateMs2(const size_t minPeaks):minPeaks(minPeaks) {}
     ~AnnotateMs2() = default;
     void createQueryList(std::vector<std::string> variableId, std::vector<std::string> ms2Id,
       std::vector<float> ms2Mz, std::vector<float> ms2Rt, Rcpp::List ms2Spectra);
@@ -22,7 +22,7 @@ public:
 private:
     std::vector<Query> queryList;
     std::vector<Reference> referenceList;
-    int minPeaks = 0;
+    size_t minPeaks = 0;
 };
 
 #endif //ANNOTATEMS2
